requirements.txt
apipkg==1.4
appdirs==1.4.3
arrow==0.10.0
asn1crypto==0.22.0
CacheControl==0.11.7
cffi==1.10.0
coverage==4.2
cryptography==1.8.1
Django==1.11.1
django-csp==3.1
django-debug-toolbar==1.6
django-environ==0.4.1
django-model-utils==2.6
execnet==1.4.1
facepy==1.0.8
funcy==1.7.3
gunicorn==19.6.0
honcho==1.0.1
html2text==2016.9.19
idna==2.5
joblib==0.11
newrelic==2.86.2.68
oauthlib==2.0.1
packaging==16.8
pep8==1.7.0
py==1.4.33
pycparser==2.17
pyOpenSSL==17.0.0
pyparsing==2.2.0
pysolr==3.6.0
pytest==3.0.5
pytest-cache==1.0
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-env==0.6.0
pytest-pep8==1.0.6
pytest-rerunfailures==2.1.0
pytest-xdist==1.15.0
python-dateutil==2.6.0
pytz==2016.10
requests==2.14.2
requests-oauthlib==0.7.0
selenium==3.0.2
six==1.10.0
sqlparse==0.2.3
watson-developer-cloud==0.26.0
whitenoise==3.2.2
